Implement mark as read logic for notifications Move some components into components file
