+++ src/components/NotificationsButton/components/NotificationsSider/components.tsx
@@ -1,8 +1,15 @@
 import Button from '@frakt/components/Button';
 import { GoBackButton } from '@frakt/components/GoBackButton';
+import { Loader } from '@frakt/components/Loader';
 import { TextInput } from '@frakt/components/TextInput';
-import { useUserNotifications } from '@frakt/hooks';
+import { UserAvatar } from '@frakt/components/UserAvatar';
+import { useUserInfo, useUserNotifications } from '@frakt/hooks';
+import { Alert } from '@frakt/iconsNew/Alert';
+import { BellSlash } from '@frakt/iconsNew/BellSlash';
 import { Settings } from '@frakt/iconsNew/Settings';
+import { getDiscordUri } from '@frakt/utils';
+import { shortenAddress } from '@frakt/utils/solanaUtils';
+import { useWallet } from '@solana/wallet-adapter-react';
 import classNames from 'classnames';
 import { FC, useState } from 'react';
 import { ContentType } from './constants';
@@ -99,6 +106,80 @@ export const SetUpContent: FC = () => {
       <Button type="secondary" onClick={onSubmit}>
         Save
       </Button>
+    </div>
+  );
+};
+
+export const NoNotificationsContent: FC = () => {
+  return (
+    <div className={classNames(styles.content, styles.noNotifications)}>
+      <BellSlash />
+      <p className={styles.noNotificationsText}>
+        You have no new notifications
+      </p>
+    </div>
+  );
+};
+
+export const LoadingContent: FC = () => {
+  return (
+    <div className={classNames(styles.content, styles.contentCentered)}>
+      <Loader size="large" />
+    </div>
+  );
+};
+
+export const SettingsContent: FC = () => {
+  const { publicKey } = useWallet();
+
+  const { data, isDiscordConnected, removeUserInfo } = useUserInfo();
+
+  const linkButtonHanlder = async () => {
+    if (isDiscordConnected) {
+      await removeUserInfo();
+      return;
+    }
+
+    window.location.href = getDiscordUri(publicKey);
+  };
+
+  return (
+    <div className={classNames(styles.content)}>
+      <h2 className={styles.contentTitle}>Channels</h2>
+
+      <p className={styles.discordSettingsLabel}>Discord</p>
+      <div className={styles.discordSettings}>
+        <UserAvatar
+          className={styles.discordSettingsAvatar}
+          imageUrl={data?.avatarUrl}
+        />
+        <div className={styles.discordSettingsUserInfo}>
+          <p className={styles.discordSettingsUserName}>
+            {data ? shortenAddress(publicKey.toBase58()) : 'Username'}
+          </p>
+          <Button
+            className={styles.discordSettingsBtn}
+            onClick={linkButtonHanlder}
+          >
+            {isDiscordConnected ? 'Unlink' : 'Link'}
+          </Button>
+        </div>
+      </div>
+      <div className={styles.discordAlert}>
+        <Alert />
+        <p>
+          Please note that you should be a member of our{' '}
+          <a
+            href={process.env.FRAKT_DISCORD_SERVER}
+            target="_blank"
+            rel="noopener noreferrer"
+            className={styles.discordLink}
+          >
+            server
+          </a>{' '}
+          to receive notifications
+        </p>
+      </div>
     </div>
   );
 };
